# openapi.yaml
openapi: 3.1.0
info:
  title: Colette API
  description: Public REST API for the Colette app. Supports OAuth 2.0 and API key authentication.
  license:
    name: MIT
  version: 0.1.0
servers:
- url: /api
paths:
  /apiKeys:
    get:
      tags:
      - API Keys
      description: List user API keys
      operationId: listApiKeys
      responses:
        '200':
          description: Paginated list of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_ApiKey'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - API Keys
      description: Create an API key
      operationId: createApiKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
        required: true
      responses:
        '201':
          description: Created API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /apiKeys/{id}:
    get:
      tags:
      - API Keys
      description: Get an API key by ID
      operationId: getApiKey
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: API key by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - API Keys
      description: Delete an API key by ID
      operationId: deleteApiKey
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Successfully deleted API key
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - API Keys
      description: Update an API key by ID
      operationId: updateApiKey
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyUpdate'
        required: true
      responses:
        '200':
          description: Updated API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /bookmarks:
    get:
      tags:
      - Bookmarks
      description: List user bookmarks
      operationId: listBookmarks
      parameters:
      - name: collectionId
        in: query
        description: Filter by the ID of a collection whose filters may apply to the bookmark
        required: false
        schema:
          type: string
          format: uuid
      - name: filterByTags
        in: query
        description: Whether to filter by tags linked to the bookmark
        required: false
        schema:
          type: boolean
      - name: tag[]
        in: query
        description: Filter by the IDs of the tags linked to the bookmark
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: cursor
        in: query
        description: Pagination cursor
        required: false
        schema:
          type: string
      - name: withTags
        in: query
        description: Whether to include the tags linked to the bookmark
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Paginated list of bookmarks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_BookmarkDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Bookmarks
      description: Add a bookmark
      operationId: createBookmark
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkCreate'
        required: true
      responses:
        '201':
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Bookmark already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /bookmarks/{id}:
    get:
      tags:
      - Bookmarks
      description: Get a bookmark by ID
      operationId: getBookmark
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      - name: withTags
        in: query
        description: Whether to include the tags linked to the bookmark
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Bookmark by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - Bookmarks
      description: Delete a bookmark by ID
      operationId: deleteBookmark
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Successfully deleted bookmark
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Bookmarks
      description: Update a bookmark by ID
      operationId: updateBookmark
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkUpdate'
        required: true
      responses:
        '200':
          description: Updated bookmark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          type: string
    ApiErrorCode:
      type: string
      enum:
      - NOT_AUTHENTICATED
      - NOT_AUTHORIZED
      - NOT_FOUND
      - ALREADY_EXISTS
      - VALIDATION
      - BAD_GATEWAY
      - UNKNOWN
    ApiKey:
      type: object
      description: Long-lived token linked to a user's account. Useful for third party client apps to access a user's data. The full value is returned only once, on creation.
      required:
      - id
      - title
      - preview
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the API key
        title:
          type: string
          description: Human-readable name of the API key
        preview:
          type: string
          description: Partial view of the API key value for identification purposes
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the API key was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the API key was last modified
    ApiKeyCreate:
      type: object
      description: Data to create a new API key
      required:
      - title
      properties:
        title:
          type: string
          description: Human-readable name for the API key to create, cannot be empty
          minLength: 1
    ApiKeyCreated:
      type: object
      description: Newly created API key, containing the full value. This value must be saved in a safe location, as subsequent GET requests will only show a preview.
      required:
      - id
      - value
      - title
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the new API key
        value:
          type: string
          description: Full value of the new API key
        title:
          type: string
          description: Human-readable name of the new API key
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the API key was created
    ApiKeyUpdate:
      type: object
      description: Details regarding the existing API key to update
      properties:
        title:
          type: string
          description: Human-readable name for the API key to update, cannot be empty
          minLength: 1
    Bookmark:
      type: object
      description: Bookmark to a webpage
      required:
      - id
      - link
      - title
      - thumbnailUrl
      - publishedAt
      - author
      - archivedUrl
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the bookmark
        link:
          type: string
          format: uri
          description: URL of the webpage the bookmark links to
        title:
          type: string
          description: Title of the bookmark
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the bookmark
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author of the bookmark
        archivedUrl:
          type:
          - string
          - 'null'
          format: uri
          description: URL of the archived version of the bookmark's thumbnail
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the bookmark was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the bookmark was modified
    BookmarkCreate:
      type: object
      description: Data to create a new bookmark
      required:
      - url
      - title
      properties:
        url:
          type: string
          format: uri
          description: URL of the webpage the bookmark links to
        title:
          type: string
          description: Human-readable name for the new bookmark, cannot be empty
          minLength: 1
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the new bookmark, will be archived
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author for the new bookmark, cannot be empty
          minLength: 1
...
    BookmarkDetails:
      type: object
      description: Extended details of a bookmark
      required:
      - bookmark
      properties:
        bookmark:
          $ref: '#/components/schemas/Bookmark'
          description: Bookmark itself, always present
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Linked tags, present if requested
...
    BookmarkUpdate:
      type: object
      description: Updates to make to an existing bookmark
      properties:
        title:
          type: string
          description: Human-readable name for the bookmark to update, cannot be empty
          minLength: 1
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the bookmark to update, will be archived
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author of the bookmark to update, cannot be empty
          minLength: 1
...
    Paginated_ApiKey:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: Long-lived token linked to a user's account. Useful for third party client apps to access a user's data. The full value is returned only once, on creation.
            required:
            - id
            - title
            - preview
            - createdAt
            - updatedAt
            properties:
              id:
                type: string
                format: uuid
                description: Unique identifier of the API key
              title:
                type: string
                description: Human-readable name of the API key
              preview:
                type: string
                description: Partial view of the API key value for identification purposes
              createdAt:
                type: string
                format: date-time
                description: Timestamp at which the API key was created
              updatedAt:
                type: string
                format: date-time
                description: Timestamp at which the API key was last modified
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
    Paginated_BookmarkDetails:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: Extended details of a bookmark
            required:
            - bookmark
            properties:
              bookmark:
                $ref: '#/components/schemas/Bookmark'
                description: Bookmark itself, always present
              tags:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
                description: Linked tags, present if requested
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
...
    Tag:
      type: object
      description: Tag that can be attached to subscriptions and bookmarks
      required:
      - id
      - title
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the tag
        title:
          type: string
          description: Human-readable name of the tag, unique per user
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was last modified
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []